requirements.txt
Django==2.1.7
django-ckeditor==5.6.1
django-js-asset==1.2.2
pytz==2018.9
